==> atp/1. Introduction.ipynb <==
# 1. Introduction

In this project, we're going to use the tennis match dataset available at: https://www.kaggle.com/hakeem/atp-and-wta-tennis-data to build a machine learning model that calculates the real chances of each player to win a match.
The goal is to use these chances to compare them against the available odds for each player in any given match, to extract value bets and see if there's some room to make an immaginary profit.


In this notebook we're going to load the original csv file with tennis matches data and we're going to explore it, to see how consistent is the data within the file.

We're going to import pandas as the library to manipulate csv files and DataFrame objects:

import pandas as pd

Let's load the 'Data.csv' file downloaded from https://www.kaggle.com/hakeem/atp-and-wta-tennis-data and stored in the csv directory within this project:

df = pd.read_csv("csv/df_atp.csv", engine='python')

Let's take a short look at the data to see if we imported it correcly:

df.head()

How many matches we have in this dataset?

df.shape

We have as many as 54908 matches and 55 attributes for each match!

Let's display the dataframe columns and their types:

df.dtypes

# Problems of this Dataset

As we can see, there are already some things that don't look very promising.
For example, the rankings of the winner player and the loser one are not of type 'int' but 'object', and the 'Date' attribute is not a date, but an 'object' too.

Moreover, in this file we have the information about who wins each match already classified in the appropriate column ('Winner'), but that's not optimal since we're going to build a classifier model that will have to learn from the attributes of each match who's more likely to win and to lose.
In fact, if the winner of the match always appears in the same column, that's not going to help our model.

We're going to correct these issues, and many more, in the PreProcessing notebook.

# Next Steps

Now that we have looked at the dataset we can explore our next steps in order to achive our goal: building a model that predicts the real chances of each player to win a match, and finds value bets to beat the bookmakers (or at least tries to)!

So, the next step is going be pre-processing this original csv file, in order to remove the features we're not interested in, and to correct the issues that we will find in the data.

Then, using the fixed data, for each match we will 'build' new features from scratch, to help our prediction model make a better job.

Next, we're going to build some machine learning models to learn from these attributes and to make a prediction about who's more likely to win a given match. And we'll compare the models in order to pick up the best one.

Finally, we will focus on the betting aspect.
In fact, we'll use the model previously built to calculate the chances of each player to win. 
Once we'll have these chances, we're going to look at the odds available for each match, to extract (if possible) value bets, i.e. bets where the odds offered by a bookmaker are higher than the odds based on the chances of a player to win the match as calculated by our model.
