==> atp/2. PreProcessing.ipynb <==
# 2. PreProcessing

In this notebook, we're going to load the original csv file already described in the Introduction notebook, and look for data to clean or fix.
Finally, we'll save a new csv file, containing the preprocessed data, to be used in the next steps.

Let's import the pandas library to work on csv files and dataframes:

import pandas as pd

df = pd.read_csv("csv/df_atp.csv", engine='python')

Let's take a look at the data

df.head()

Since we're not going to use every column in the dataset, first we're going to list all the columns, and then we'll delete the unnecessary ones.

list(df.columns)

df = df.drop(['Unnamed: 0','ATP','Tournament', 'Location', 'Best of', 'W1', 'L1', 'W2', 'L2', 'W3', 'L3', 'W4', 'L4', 'W5', 'L5', 'Wsets', 
              'Lsets', 'Comment','CBW', 'CBL', 'GBW', 'GBL', 'IWW', 'IWL', 'SBW', 'SBL',
              'B&WW', 'B&WL', 'EXW', 'EXL', 'WPts',
              'LPts', 'UBW', 'UBL', 'LBW', 'LBL', 'SJW', 'SJL'], axis=1)

df.head()

Now, as we saw in the Introduction notebook, the 'Date' attribute is not a date, but it has 'object' type.
Let's convert these objects into dates:

df['Date'] = pd.to_datetime(df['Date'])

df['Date']

Now the 'Date' column contains dates only.

Next, if we take a look at 'Series' column, we can see that it contains old-labeled tournament series (i.e. International Gold for ATP500, etc.)

df['Series'].unique()

So, let's rename those values to make sure that each torunament category is different from the others.

df['Series'] = df['Series'].str.replace('International$','ATP250', regex=True)
df['Series'] = df['Series'].str.replace('International Series$','ATP250', regex=True)
df['Series'] = df['Series'].str.replace('International Gold$','ATP500', regex=True)
df['Series'] = df['Series'].str.replace('Masters$','Masters 1000', regex=True)

df['Series'].unique()

Now let's explore the 'WRank' and 'LRank' attributes:

#let's take the union of the sets of 'WRank' and 'LRank' values and visualize them
set(df["WRank"].unique()) | set(df['LRank'].unique())

As we can see, the values above are not numbers, but objects, and there are values like 'NR' or 'nan' (look at the end).

Let's change the types of columns WRank and LRank from object to int.
In order to do this, we firstly change their type to 'str', we replace values like 'NR' or 'nan' with the standard value of 500, and finally we cast the types to int.

df["WRank"] = df["WRank"].astype('str')
df["LRank"] = df["LRank"].astype('str')

df["LRank"] = df["LRank"].str.replace("NR$","500",regex=True)
df["WRank"] = df["WRank"].str.replace("NR$","500",regex=True)
df["LRank"] = df["LRank"].str.replace("nan$","500",regex=True)
df["WRank"] = df["WRank"].str.replace("nan$","500",regex=True)
df["WRank"] = df["WRank"].str.replace("\\.0$","",regex=True)
df["LRank"] = df["LRank"].str.replace("\\.0$","",regex=True)

df['WRank'] = df['WRank'].astype('int')
df['LRank'] = df['LRank'].astype('int')

df.head()

Another problem with this dataset, as we stated in the Introduction notebook, is that the players are already classified as "Winner" and "Loser".
Since we will train a model to identify who's more likely to win given a list of match attributes, we want to fix this issue, otherwise the model will figure out that columns relative to the winner are more 'important' than others to detect who'll win the match.

That's why we decided to:
- change the name of the columns "Winner" and "Loser" in "Player 0" and "Player 1".
- for half of the records, we'll swap "Player 1" attributes with "Player 0" attributes and viceversa, in this way we'll distribute the winners equally between "Player 0" and "Player 1". 
- create a new column "Won" that will contain 0 if Player 0 has won the match or 1 otherwise.

df = df.rename({"Winner":"Player 0", "Loser":"Player 1", 
                "WRank":"Pl0_Rank", "LRank":"Pl1_Rank",
                "MaxW":"Max_Pl0", "MaxL":"Max_Pl1", "AvgW":"Avg_Pl0", "AvgL":"Avg_Pl1",
                "B365W":"B365_Pl0","B365L":"B365_Pl1","PSW":"PS_Pl0","PSL":"PS_Pl1"
               }, axis=1)

for ix, row in df.iterrows():
    if ix % 2 == 0: 
        #swap half of the records
        old_p0 = row['Player 0']
        old_p1 = row['Player 1']
        old_p0_rank = row['Pl0_Rank']
        old_p1_rank = row['Pl1_Rank']
        old_max_p0 = row['Max_Pl0']
        old_max_p1 = row['Max_Pl1']
        old_avg_p0 = row['Avg_Pl0']
        old_avg_p1 = row['Avg_Pl1']
        old_b365_p0 = row['B365_Pl0']
        old_b365_p1 = row['B365_Pl1']
        old_ps_p0 = row['PS_Pl0']
        old_ps_p1 = row['PS_Pl1']

        df.at[ix,'Player 0'] = old_p1
        df.at[ix,'Player 1'] = old_p0
        df.at[ix,'Pl0_Rank'] = old_p1_rank
        df.at[ix,'Pl1_Rank'] = old_p0_rank
        df.at[ix,'Max_Pl0'] = old_max_p1
        df.at[ix,'Max_Pl1'] = old_max_p0
        df.at[ix,'Avg_Pl0'] = old_avg_p1
        df.at[ix,'Avg_Pl1'] = old_avg_p0
        df.at[ix,'B365_Pl0'] = old_b365_p1
        df.at[ix,'B365_Pl1'] = old_b365_p0
        df.at[ix,'PS_Pl0'] = old_ps_p1
        df.at[ix,'PS_Pl1'] = old_ps_p0

        df.at[ix,'Won'] = 1
    else:
        #don't swap these records, but just mark as winner Player 0
        df.at[ix,'Won'] = 0

df.head()

Next, we are going to create dummy variables dataframe for the columns: 'Court', 'Surface', 'Round' and 'Series'.

Then we are going to add these dummy variables dataframes to the original one.

court_dummies = pd.get_dummies(df['Court'])
court_dummies.head()

surface_dummies = pd.get_dummies(df['Surface'])
surface_dummies.head()

round_dummies = pd.get_dummies(df['Round'])
round_dummies.head()

series_dummies = pd.get_dummies(df['Series'])
series_dummies.head()

Now we're going to merge all these dataframes and append those to the original one:

df = pd.concat([df, court_dummies, surface_dummies, series_dummies, round_dummies], axis=1)
df.head()

Now we're going to the delete the 'Court', 'Surface', 'Round' and 'Series' columns since we already created the respective dummy columns:

df.drop(['Court', 'Surface', 'Round', 'Series'], axis=1, inplace = True)

df.head()

Finally, the preprocessing stage is concluded!

Now, we're going to save the preprocessed dataframe in a csv file, so that the next notebook will use this dataframe directly after loading it!

df.to_csv("csv/Preprocessed_Data.csv", index=False)
