==> atp/5. Value Betting Calculator.ipynb <==
# 5. Value Betting Calculator

In this notebook we'll make use of our match dataset and machine learning model we've built in the last stage, to predict the chances for each player to win in any given match.

We'll then compare these chances with the odds available for those matches, and build a Dataframe where each row will contain information regarding which player we would have to bet on!

Let's import the libraries needed for this notebook:

import pandas as pd
import pickle

Now, we'll load the model trained in last notebook:

model = pickle.load(open("logistic_regression.lr", 'rb'))

Now, let's open the Coefficients dataframe saved in the last notebook, and let's create a dictionary object based on the dataframe, where the key is going to be the attribute name and the value will be the attribute's coefficient:

coefs_df = pd.read_csv("csv/Coefficients.csv")

coefs_df.head()

coefs_map = {}
for ix, row in coefs_df.iterrows():

    attribute = row['Column']
    attribute_coef = row['Coef']

    coefs_map[attribute] = attribute_coef

coefs_map

Let's make a quick analysis on the greatest indicators for player 1 to win a match against player 0.
Let's see which attributes matter the most:

coefs_df[abs(coefs_df['Coef']) > 0.1].sort_values('Coef', ascending = False)

We can see that the most important features are:
- Player performance on specific surface
- Player performance against similar ranked opponents
- The opponent's form
- The rank index
- The players recent form

Let's load our matches Dataset

df = pd.read_csv("csv/FeatureCalculated_Data.csv")

Now our goal is to build another Dataframe, containing the informations about the the players, who wins each match, the average and maximum odds available for each player, and we'll use our loaded model to calculate the predicted chances to win for each player:

betting_data_df = df[(df["Avg_Pl0"] > 1) & (df["Avg_Pl1"] > 1)].copy()

inputs = betting_data_df.iloc[:, 14:]

predicted_outputs = model.predict_proba(inputs)

index = 0
for ix, row in betting_data_df.iterrows():
    prob_pl0 = predicted_outputs[index][0] 
    prob_pl1 = predicted_outputs[index][1] 
    betting_data_df.loc[ix, 'Pl0 %'] = prob_pl0
    betting_data_df.loc[ix, 'Pl1 %'] = prob_pl1
    index += 1

betting_data_df.head()

Let's define a helper funxtion that takes a change in input and returns the 'fair' odds for that chance value.

def right_odds(chance):
    return round(1/chance,2)

Now, we have everything we need to make decisions based on which player, for each match, we get the most value betting on!
In fact, next we're going to append to this betting dataframe another 4 columns.

The fair odds for both player 0 and player 1, based on the chances of each player to win according to the model, and the value that we would get betting on each player at the average odds available!

We use the average odds to determine value in order to get a more realistic outcome.

for ix, row in betting_data_df.iterrows():
    right_odds_p0 = right_odds(row['Pl0 %'])
    right_odds_p1 = right_odds(row['Pl1 %'])
    bet_value_p0 = (row['Avg_Pl0'] - right_odds_p0)/max(right_odds_p0,row['Avg_Pl0'])
    bet_value_p1 = (row['Avg_Pl1'] - right_odds_p1)/max(right_odds_p1,row['Avg_Pl1'])
    betting_data_df.loc[ix, 'Fair odds for Pl0'] = right_odds_p0
    betting_data_df.loc[ix, 'Fair odds for Pl1'] = right_odds_p1
    betting_data_df.loc[ix, 'Bet on Pl0 Value'] = bet_value_p0
    betting_data_df.loc[ix, 'Bet on Pl1 Value'] = bet_value_p1

betting_data_df.head()

As we can see, there are bets for which nor betting on Player 0 nor on Player 1 would be a good choice (Bet on Pl0 Value and Bet on Pl1 Value are both negative).
Since there's no point in betting where we don't think to have an edge, we will remove these records from the dataframe:

value_betting_data_df = betting_data_df[(betting_data_df["Bet on Pl0 Value"] > 0) | (betting_data_df["Bet on Pl1 Value"] > 0)].copy()

value_betting_data_df.head()

Let's rename some of the columns to help the readability of the dataframe

value_betting_data_df.rename(columns={"Avg_Pl1":"Pl1 Avg odds", "Avg_Pl0":"Pl0 Avg odds", "Max_Pl1":"Pl1 Max odds", 
                                "Max_Pl0":"Pl0 Max odds", "Won":"Winner"},inplace = True)

Let's save this dataframe!

value_betting_data_df.to_csv("csv/Betting_Data.csv", index=False)
